+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - none
 
 # v2025.2.1-beta
+- sqlmath - Update sql-function SHA256() to use tinycrypt.
 - jslint - Update shell-function shDirHttplinkValidate() with pragma <\!!--novalidate--\>.
 - ci - Upgrade cpplint.py used in ci to v2.0.0.
 - sqlite - Rename custom-sql-function MEDIAN() to MEDIAN2().